bin: don't "hide" Orocos::NotFound once we got a reference on the TaskContext
The current implementation of bin/orocos_state was rescuing Orocos::NotFound
in the pretty printing. That would hide bugs in which the pretty printing
code tries to get a reference which does not exist.

Nonetheless, it is important to rescue it around the call to TaskContext.get,
since it is possible that there is a dangling object in the nameserver.