cleanup CORBA low-level layer
The manipulation of the CORBA layer is done in methods
of the Orocos::CORBA module, and are defined either in
ext/corba.{cc,hh} or in orocos/corba.rb. New methods to
manipulate OmniORB's timeout have been added. Moreover,
the initialization of the CORBA layer is done in a separate
CORBA.init method, allowing later to load the framework
without CORBA.

A new tool, orocos_ns_cleanup has been added. This tool
removes all dead references from the nameserver.