link with orogen
After this commit, orocos.rb require that the processes it manages have been
generated by orogen. It then can link a TaskContext object to its orogen model.

If a task is taken directly from TaskContext.get, the associated process is
searched and -- if found -- associated.

This fixes some issues with #spawn, because orocos.rb now waits for all tasks of
a given process to be spawned -- and does not rely on the name to do so anymore.

Finally, a log_all_ports method has been added, which searches for the Logger
object that orogen's StaticDeployment#add_default_logger, and logs all output
ports in the process.