fix finding task models for tasks whose process is unknown
Until now, we were relying on the availability of the task's
process to find the task's model -- which is needed for most
of orocos.rb functionality. This obviously fails miserably
for remote tasks.

With the last commit, it had been changed so that we use the
orogen-generated getTaskModel method to get the task's model
and then try to find it on the local machine.

This commit cleans up that functionality, makes it way faster
by chaching the list of available task models and fixes it.