roby: finally got the data source handling right (hopefully)
Now, everything is centered on data sources. They can be declared
as root data sources or slave ones. Devices are handled
as normal data sources, simply being of a more specialized type
(DeviceDriver). Subdevices are data sources that are slaves of a
given device.

At the model level, we can now represent tasks that handle any
combination of sources/devices. The integration in the instanciation
chain is not yet done though.