roby: silently overload matching main data sources if :as is not given
In the following situation:

  TaskModel.data_source Camera
  SubTaskModel.data_source FirewireCamera

with SubTaskModel < TaskModel, I find it less surprising to
have the FirewireCamera data source *overload* the Camera
data source of the parent model, instead of creating a new
model. So, if :as is not explicitely given, do overload.
Using :as will give access to the other behaviour.