roby: make can_merge? more conservative on the task model
Two side of our representation are fighting each other:
 1. the plan model which represents what is _required_ by
    the parent tasks.
 2. the task allocation scheme that can choose a specific
    task model for whatever reason (in our case, because the
    use asked him to).

Until now, the merge algorithm was only considering (1) and
not (2), and was therefore overriding the task allocation
scheme.

From now, the merge algorithm should be made more conservative.
If changes to the plan are needed to optimize/change the
configuration, it is not its job.