roby: change I/O compatibility check
During the merge, the algorithm was only checking that
the two tasks have the *same* inputs, but was failing
to recognize a merge if the they have inputs that do
not overlap (i.e. ports in target_task that are unused
in +task+).

Change the test, and move it to can_merge? so that it
applies to abstract task allocation as well