roby: in #link_to_busses, search for a generic com port only if we need one
We were before searching for generic I/O ports to the communication
bus, even though the required interfaces were already properly
connected to the bus. There is no need for that, and it can even
be harmful if some interfaces that the module support are to be
ignored by the user (i.e. if there are some ports that can be
connected to the bus, but that should be ignored)