roby: really integrate the plugin into Roby
The main change is that the setup stage is now split between
a require_models and a setup method. The first one loads the
model files from tasks/ while the second one initializes
the singleton constants and attributes.

The main functionality change is that whenever a project is
loaded, a corresponding file is searched for in tasks/orocos,
and loaded if it exists.