implement a clean shutdown of processes
after this commit, #kill will by default stop all tasks,
deconfigure the ones that need to be configured, and disconnect
all ports. Then, it calls ServiceInterface.requestShutdown
to cleanly shutdown the process.

If the "clean" method fails for whatever reason, it will resort
to sending a signal again, and will report that it is doing so.
Moreover, it will report if the process died an unexpected way.

Finally, it is possible to optionally use signals instead of
shutting down the clean way.