roby: fix connection autodetection for needs_(buffered|data)_connection
This fixes a bug: when needs_buffered_connection was set, a policy
with :type => :data was chosen anyway. Only needs_reliable_connection
was making the code choose a buffered connection.

Now, needs_buffered_connection without needs_reliable_connection
simply returns a buffer of size 1. The rationale is that, if the
user wants a reliable connection, then he should set
needs_reliable_connection.

Note that this is meant as a workaround of the RTT 1.x behaviour,
where data and buffer ports have different properties from the
P.O.V. of the target's code (i.e. not only on the system's behaviour).
It will be fixed by RTT 2.0 ports.