roby: #specialize should not cross-apply specializations if they are not applicable
This fixes cases where the same composition is specialized on
two different component tasks. Since these tasks are not related
to each other, it is not possible to apply the specialization to
each other (but it is possible to create the two specializations from
the composition).

This commit makes #specialize check if a specialization applies to
already existing specializations before calling #specialize