roby: improve handling of composition constraints requirements
We now mark as abstract every specialization that does not meet
the requirements. Moreover, when possible, we also remove
such specializations.

The limit, right now, is that we can't remove specializations that
have children as the specialization code does not support having
multiple specialized models for the same child at the same level
of the composition hierarchy.