roby: resolve explicit selection early, and make the selection of services more robust
Until now, the resolution of names given to InstanciatedComponent#use
was done late in the instanciation process. This was making the whole
process harder (as more information had to be propagated) and was
really not flexible for data service selection.

Move the resolution of explicit selections to
InstanciatedComponent#instanciate, and allow to explicitely select
any service, by using the new ProvidedDataService data structure.