roby: rework the merging algorithm
The new algorithm works directly on the merge graph, as it should
have been. It removes some (if not all) of the problems in ambiguity
resolution in the old algorithm.

Ambiguity resolution is now a lot more robust, and will take into
account process names (matching them to device names) and process
locality (will prefer merging tasks that reduce the distance in the
data flow)