roby: allow to specify if an instance *should be* or *has been* removed
This is the first part of the effort of creating a seamless bridge
between the plan management and the orocos engine.

We now have the task selection tasks. They add new tasks to the orocos
network in a nice way. However, there is still no way to propagate
normal Roby GC towards the engine.

When the GC removes a task, it means that the task is not needed anymore.
So, it should be removed from the engine's requirement set. However,
it is not a *request* to remove a task. The task *has been removed already*

Therefore, we treat it differently: the task is first removed normally.
If #resolve works fine, then everything stops there. However, if #resolve
fails, then a second resolve is run when only these "already removed
tasks" are handled, to update the engine's data structures.