roby: fix shutdown of deployments
Commits related to working around omniorb bug have introduced a bug in
the deployment exit process. Indeed, tasks' state was not updated
anymore as soon as the deployment was marked as finishing. However, in a
first stage, finishing deployments are waiting for their tasks to
finish...

Introduce a new predicate, "ready_to_die?", that is set on finishing
deployments as soon as all tasks are finished.