roby: make direct_merge_mappings compute merges the other way around
Previously, the method was iterating over the provided task set
and checked if the considered task could be used to replace other
tasks.

This is inefficient, and reduces the applicability of merges.

Inefficient, as some checks (for instance, is the target task a proxy ?,
does the target task have an execution agent ?) are done in the inner
loop instead of the outer loop.

Reduces the applicability as we use queries to check for tasks, but
the orocos/roby plugins does a bit of tweaking internally for data
service proxies. It actually means that one was never able to replace
data service proxies (!), which was the reason why the task allocation
pass was necessary.