roby: during composition instanciation, make a difference between explicit and implicit selection
We now make a difference between the models selected because they
are listed in the explicit selection map, and the ones that we add
on the fly to represent the "holes" in the selection.

This is because some selection algorithms -- as for instance
specialization selection -- don't want to override user selection.
As such, a specialization will *not* be selected if one of the children
does not match the explicit selection. It will, however, be selected
if said children are not explicitely selected.