do *not* initialize State.config.task_name with default values
It is hard, and actually not-so desirable, to always initialize properties in the
orogen file. We therefore want to be friendly with modules that initialize their
properties in the task's constructor.

So, do *not* force the writing of default values by initializing State.config.task_name

However, we will have to provide a nicer mechanism to store configuration in State.config
anyway.