modify the API to reflect for the tristate now returned by OutputPort::read
The outputport class now has a #read and a #read_new method. The first
one will return a non-nil value in both the OLD_DATA and NEW_DATA cases
while the second one will do it only in the NEW_DATA case.

Modified tests to use the read_new method when applicable, and updated
the test components.