roby: we should not define #config_type too eagerly
First of all, the config_type accessor should be at the model
level, not the instance level as it is the case right now.
Moreover, if a data service model has explicitely been created,
one should assume that its config_type has been set appropriately.
We therefore should not override it in #driver_for with the task
context's configuration type.