roby: fixes to service mappings during instanciation
The service mapping code is in charge of looking for task services
matching the data services required by the composition's children.

However, it was searching for a single service that was fullfilling
all the requirements. This is broken as one single child specification
can require orthogonal services. So, find services one by one.

Moreover, by raising exceptions, it was breaking the semantics of
#find_selected_models_and_tasks (and, more importantly, #narrow). Fix
that by adding a 'user_call' parameter as for the other methods.