roby: in Deployment#dead!, do not emit :success or :failed if a terminal event has already been emitted
It is possible to signal :stop (or another terminal event) from
an event which is itself terminal. In this case, it is expected
that the final task state is determined by the first terminal event

So, in #dead!, only emit :stop if a terminal event has already
been emitted, to avoid overriding the final task state.

It fixes a bug where the emission of internal_error was followed
by :success (!)