roby: fix merge_sort_order
We use a [boolean, boolean] truth table to return the sort order
value, but it is common in Ruby that predicate methods don't return
true/false but "anything" vs. false-or-nil. So use the !!value idiom
to test a true value.

Fix the execution agent merging order, in which we were preferring
using non-deployed tasks to deployed tasks (!)