roby: fix requiring multiple dynamic slaves on a given task model
The issue was that, the second time a dynamic slave was created,
it was created on top of the *original* model instead of on top
of the already-specialized one. In the end, it led to having only
the second one available, and the third one made everything fail
as we could not create the oroGen interface spec at all.