roby: when creating a dynamic slave, we need to overload the master service as well
The issue is that when we do

  master_srv = base_model.find_data_service("master")
  srv = submodel.require_dynamic_slave(master_srv)

then master_srv.component_model == base_model. Therefore, if
I were to look at

  master_srv.component_model.each_slave_data_service(master_srv)

then the new slave would not exist. Idem for static slaves added
in subclasses of other component models.

This commit creates an overloaded version of master_srv in
specialized_model whenever a slave is created there. It also
creates ProvidedDataService#each_slave and adds an argument
to ProvidedDataService#each_(input|output)_port that allows
to enumerate both the master service's ports and its slaves