better interaction with the ORBInitRef envvar
There was an inconsistency in the handling of ORBInitRef:
 * the ruby extension would connect to ORBInitRef first and to the value
   in CORBA.name_service second
 * the processes would always be started with the value in CORA.name_service
   as name server

This commit changes that behaviour: if ORBInitRef is set and CORBA.name_service
is not explicitely set, prefer ORBInitRef. Otherwise, override the value in
ORBInitRef by the value in the script.

The behaviour of connecting to the local name server if neither ORBInitRef
nor CORBA.name_service are set is kept