roby: in added_child_object, when listing modified tasks, add both parent AND child
The connection manager uses a set of modified tasks that gets
updated by the tasks' #added_child_object and #removed_child_object
hooks.

However, these hooks were only adding the parent of the added/removed
connection to the set of modified tasks. In effect, it meant that
a connection A -> B was only added or removed properly if - by any
chance - B's output is also modified at the same cycle.

Fix by adding both parent and child to modified_tasks