roby: fix handling of default value for :conf
The general issue is that the merging code cannot handle
"static" default values like :conf

So, we do *not* set a default value for :conf so that the
merging code can merge tasks on which :conf is set with tasks
on which it is not set (this is critical).

However, in order for the network to run, :conf needs to be
set at the very end. So, set it to ['default'] at the end
of #resolve on all tasks on which it is not set.