roby: fix event emissions related to exception states
This had been broken in 6071fd44097275e1e27bbfe21b31f98006483393,
where the call to reset_exceptions has been removed.

Beforehand, reset_exception was causing a transition to STOPPED
or PRE_OPERATIONAL, which was handled in a different part of
handle_state_changes. Removing that, we need to emit the exception
event directly.