roby: resolve selections recursively
This commit solves the common issue of using #use to configure
a composition, and then use this (configured) composition to
provide some services, as e.g.

  use Srv::Pose => Cmp::PoseEstimator
  use Cmp::PoseEstimator => Cmp::PoseEstimator.use_conf('bla')

This was currently not working as the two use statements were
not combined to resolve to the (usually expected)

  use Srv::Pose => Cmp::PoseEstimator.use_conf('bla')
  use Cmp::PoseEstimator => Cmp::PoseEstimator.use_conf('bla')

This commit fixes that by resolving the selections recursively.
Moreover, the selection resolution code has been refactored to get
rid of the "default selection" entry, i.e.

  use Model1, Model2

is now translated to a proper

  selected_model => Model1, selected_model2 => Model2

mapping.