core: fix the dataflow management code to trigger if new connections are added between tasks that are already connected
The issue was that the connection updates were triggered only
when relations in the DataFlow graphs got removed or added.

However, if one adds a connection between tasks that are already
connected through other ports, then these hooks don't get called.
Watch for the updated_info hook to trigger the update here as well