roby: fix wrong place for calling validate_generated_network
The validation was happening *after* we started deploying the tasks.
This is wrong, as it is supposed to validate the "plain" network
(i.e. not deployed at all).

It was triggering wrong reported failed task allocations: the
placeholder tasks were getting into the transaction and would trigger
the validity check on abstraction.