roby: fix handling of possible cycles in the internal loop
The merge is incremental. To avoid complex calculations because of
dataflow loops, we first resolve all simple merges and then resolve
the cycles.

However, the possible_cycles array was not updated using the merges
that happened. It could therefore contain cycles between tasks that
are not valid anymore (because they got merged in between)

To avoid this, we now re-add the possible cycles, updated by the
merges, to the candidates so that they are part of the next pass.