roby: replace the merge cycle breaking algorithm
The old algorithm was local and ... let's put it frankly ... sucked
in all but the simplest cases.

The new one performs finds, for each of the tasks that are part of
a cycle, the task from which the most other tasks can be reached. All
back pointers in the resulting tree are removed to leave only direct
links.