roby: major cleanup of the update of #task on instances and devices
This introduces a global merge result graph in the NetworkMergeSolver.
This graph allows to find out, for any task that was once in the plan,
by which task it has been replaced so far. It fixes the mess that was
the update of instances[].task and devices[].task, and fixes a few
issues in the process