roby: fix generation of connection policies which involve transaction proxies
The dataflow dynamics information is generated in the post-deployment,
pre-finalization stage of the resolution, while the logging is configured
in a post-finalization stage. In practice, it means that the logging
was referring to transaction proxies while the dataflow dynamics was
referring to intermediate tasks.

This fixes the issue by applying merge mappings to the dataflow dynamics
information