roby: fix handling of placeholder tasks when one of the models is a task model
So far, in a lot of places, the task model was taken as-is. This is
broken as the additional services are *not* provided by the task model
(or they would have been removed from the requirement list) and therefore
the task model does not encode everything.

This was working so far as a well-formed specification does not have
any such instance left. It breaks in model browsing code for instance,
were we *do* have partial instanciations