Revert "roby: do not attemp to connect tasks that failed to be configured"
This reverts commit b58fbcd3fa91fd0041205d036685936b2c0d49d5.

This change is both unneeded and harmful.

Harmful, as the job of #compute_connection_changes is to compute the set of
connections that need to be created for the ActualDataFlow graph to match the
information in DataFlow. This change was "hiding" such information, which was
leading to some connections not being created: some non-setup tasks at the time
of #compute_connection_changes were not stored in the new set, which was
expected, and therefore "forgotten" in the update process.

Unneeded because the filtering was already done in apply_connection_changes