roby: fix handling of specializations for narrowed models
If an InstanceRequirements got narrowed on specializations (S1, S2),
but at instanciation time we needed to add S3 to the list, the new
specialization was not added.

The issue was that the instanciation was done on the narrowed model,
which currently has no specialization. Workaround the issue by always
instanciating on the root model (which means that we don't benefit
from narrowing to reduce computation costs of instanciation). In the
long run, we should compute the specialization compatibility graph and
fill in the specializations in instanciated specialized composition models.