refactor the dependency injection code
This tries to move all the code necessary to do instance selection into the
dependency injection code, to finally remove it from Engine and Composition

The new code's main interface is DependencyInjection#instance_selection_for,
which resolves a selection and returns the information necessary for
instanciation: task model, service selection and port mappings.

Moreover, DependencyInjection now normalizes the selection, making sure as soon
as possible that the user selections are valid.