clear the roles of DependencyInjection, InstanceRequirements and InstanceSelection
InstanceRequirements (which will be renamed to InstanceConfiguration
at the end) represents a configured component

DependencyInjection allows to match a desired configuration to a
selected configuration

InstanceSelection represents the necessary information to map a placeholder
task from the required configuration to the selected one, such as
service mappings (and therefore port mappings).

The goal in the end is that the engine and composition should not have
any code dealing with that kind of issues anymore.