gui: we need to explicitly enable/disable views
We are reusing the rendering objects, which are therefore always
listening to the button URIs. Therefore, all the subclasses of
ComponentNetworkBaseView will try to re-render their plan each
time an annotation is changed, which leads to having blank dataflow
displays.

This commit introduces an enable/disable pair of methods that are
called by the model browser to make sure that only one managing object
is active at a time.