merge_solver: pass the possible_cycles information across passes
The merge passes don't work on the whole plan - we try to pass
only the tasks whose merge status has changed from the previous
pass. Therefore, an item that has been found as a possible cycle
in pass X might not be found in pass X+1. Since we process the
cycles only when there are no simple merges to perform, we have
to store all found cycles until we process them.