merge_solver: replace return by next in the simple merge processing
In #apply_merge_mappings, when processing the simple merges, we were
returning as soon as an ambiguous merge is found. This is causing
proper simple merges to be ignored (as #apply_merge_mappings was
not passing through all possible simple merges). Fix that by replacing
the return by a next.