don't add running but unneeded toplevel tasks to the list of current requirements
This fixes the following situation:
  A runs
  B is pending

on event e (which is NOT a terminal event of A), A is unmarked, but not
killed (we expect the GC to do that).

Then B is started by the scheduler. Since A and B are both running,
the engine would currently try to deploy both. The intent, obviously,
is to run only B.

Moreover, we can't run the GC before doing the engine deployment,
since we want the engine to reuse as much as possible from the
current plan.