InstanceSelection: fix application of service selection in #instanciate when a component object is explicitly given
instanciate modifies #selected by selecting the target service, when
only one service is selected. This was not done when a component object
is directly given to InstanceSelection.new

Moreover, #instanciate is expected to return an object with said
selected service applied. This was also not done in this code path.