test: update test, #configure is now always called
The rationale was in the original commit

commit 43e3ce5f8e54f1ed7e4d4846e1183ce212f05ac6
Author:	Sylvain Joyeux <sylvain.joyeux@m4x.org>  Sun May  5 17:42:29 2013
Committer:	Sylvain Joyeux <sylvain.joyeux@m4x.org>  Sun May  5 17:42:29 2013

always call #configure, even if the task's configuration did not change

The configure method is user code, so we don't know what it actually
wants to do. Don't assume that because the 'conf' flag did not change,
the configure method won't do anything.