make the behaviour of Models::DataService#provides consistent when both services have the same port
The documentation says that x.provides(y) *adds* ports from y to x.
When both have the same port (same name, same type), this should
obviously lead to a collision.

Makes this the behaviour, documenting in the error message that the
solution is to add an explicit identity mapping, e.g.

   x.provides y, 'port' => 'port'