test handling of RubyProcessServer/RubyTaskContext in Syskit::Deployment
It makes sure that #orocos_task is set to the RubyTaskContext instance
and not plainly the TaskContext. It does so by delegating task resolution
to the process instance, which will allow in the future to handle
ROS nodes and log tasks the same way. Moreover, it will also allow
us to rely less on task names for resolution.