improve dynamic data service API
This puts the classes related to dynamic service classes into their
own files (as it should), and makes sure that we can determine whether
a given service has been created from a dynsrv or not (extremely useful
in #configure).

It also adds a with_dynamic_service call to the component models that
returns the specialized dynamic data service.