properly type children when in a specialization block
This fixes a bug in which specialization blocks were evaluated in a
context that depend on the other specialization that apply. Namely,
one would expect connection / export policies in a given specialization
block to refer to the specialized model (i.e. data service port names),
but they were actually applied to the model of the composition on which
all specializations have been applied.