refactor service handling in InstanceRequirements
The major change is that InstanceRequirements does not manage sets
of models anymore. It delegates that to task proxy models which, for
the occasion, have been modified to handline proxying any kind of
component model (taskcontext, component, compositions).

The main effect is to simplify all the code that modify
InstanceRequirements, as we are now guaranteed to get a single
component model. Moreover, I made sure that the IR API now allows
to access the necessary information about the IR without accessing
model or base_model directly.

It also finally clarifies the semantic of service selection in IR:
there can be only one data service selected, and this is represented
by having #base_model and #model point to this particular data service.