fix service selection in DependencyInjection when adding by-name selection
The faulty behaviour was to register the object as a selection for its
own service(s). This was failing when a service is required, but the
selected object provides a submodel of that service, e.g.

  sub_srv.provides srv
  task.provides sub_srv, :as => 'test'
  di = DI.new('child' => task)
  selection_for('child', IR.new([srv]))

was returning a selection for sub_srv, while what we need to do proper
mapping computations is a selection for srv.

In general, we now compute the mapping from the requirements to the
selection