Composition#find_required_composition_child_from_role must return a service that is faceted to the requirement
So far, the method was returing the service of the child that was
selected during instanciation. However, it is possible that the
composition was requiring a service, while the task was providing
a sub-service of it (e.g. composition needs Position while task
provides Pose). The returned value of find_required_... should in
these cases be the composition requirement (Position) and not the
task's (Pose)