fix deployment of components that have been specialized with Component#specialize
When one component got some dynamic services instanciated,
components it gets merged into will be specialized with Component#specialize,
which creates a singleton class.

The issue then was that the reconfiguration code could not instanciate
a new task of that particular model. We instanciate now using #concrete_model
instead of #model, #merge deals with specializing the new deployed task.