use selected services as a way to filter out ambiguous specialization selections
We select specializations based on the selected component models, as any other
way would make specialization selection extremely impractical

However, the issue is of ambiguous specializations. The intuitive way would be
to select in the use() statement the service that would allow to disambiguate,
as e.g.

  Composition.use('child' => ATask.as(ServiceModel))

This commit implements that particular scheme. The task model is still primarily
used to find which specializations apply, but the service is still used if the
specializations are ambiguous