fix some issues with specialization disambiguation
So far (mostly in rock-roby), we were relying on service selection
as a way to disambiguate specialization selections. However, it is
weird since it involved being able to do

  'child' => Task.as(service)

even though the child is actually a Task. This does not work in
syskit as syskit does typechecking at this level.

This still fixes the disambiguation mechanisms through service
selection, but also introduces a more generic #prefer_specializations
method on InstanceRequirements to explicitly select them.