don't cache the created specialization in SpecializationManager#instanciate
This is a workaround for a problem with applying specializations
when subclassing compositions. The issue is that we apply specializations
at the time of class creation, which means that any modifications (e.g.
overloads) applied to the subclass will *not* be seen by the composition
model created in #specialize (making it invalid)

By not registering the model, we make sure that #specialized_model
will create one at use time, i.e. after we do know the subclass model
fully