freeze delayed arguments at the end of the component network generation
At this stage of the deployment, every options / arguments should
be set so that we can compare/adapt the running network. The paradigm
in syskit is to interpret "unset argument with defaults" as "pick
whatever for this argument" (thus allowing merging). However, as
soon as we are done with the generation, we freeze the delayed
arguments so that the system designer is guaranteed to get always
the same network from the same spec (regardless of what is already
running)