do not store instances in Composition#child_selection when use() refers to a composition child
Storing instances is not robust to replacement, and is actually breaking
the contract here

Indeed, we should be able to re-instanciate using the information
stored in #child_selection, but storing an instance does not represent
that we want "this other child of the composition". It says
"this particular instance" instead.