never modify specialized models when merging tasks that have required dynamic services
The current code was calling #specialize only if the model was already
not a private specialization. However, this was causing a real class
created from #specialize to be modified because of the merge. Since
the same class could have been used to create other instances, it
can have unintended side-effects.

Make instead sure that the target task in the merge will create the
new services in its singleton class. To avoid unnecessary overhead,
do not do it if no services need to be instanciated.