fix creating submodels of a specialized composition model
The general assumption in the code is that specialized composition
models cannot be subclassed. However, this is necessary when e.g.
creating proxy task models based on specialized composition models,
and in general for consistency reasons.

Since we don't want to break the underlying assumption, #new_submodel
now creates a new composition submodel on which the same specializations
are applied.