add proper synchronization primitives for input writers
When one wants to write on an input only once, he has to wait
that the task that provides the *actual* port is running. Otherwise,
the sample might be lost (in the orogen case, *will* be lost
as orogen components clear their inputs on start)

This changes #write to return false if the writer was not ready
to be written on, and adds the #wait_until_ready task script
primitive to wait for a writer to be ready.