netgen: in #process_possible_cycles, return right away if straight merges got found
Because of merges during the incremental pass, the possible_cycles
set might contain tasks that can (now) be merged together straight
away. The set is post-processed to find those at the top of
process_possible_cycles.

However, straight merges were removed from the possible_cycles set,
and the rest of cycle resolution was actually executed. The net effect
was that the cycle resolution would invalidate any cycle containing
tasks that can be merged straight (since they are not part of
possible_cycles anymore). In itself, it could have been fine, but
since the task pairs *also* got removed from the possible_cycles set,
we would not retry merging them once the straight merges were done.

The fix is to return to the incremental pass right away as soon as
straight merges are found in process_possible_cycles