netgen: allow for more than 1-to-1 mappings for tasks in cycle resolution
The assumption in cycle resolution so far is that, for any given
cycle, one task would be merged with at most another task. This
is actually too limiting for complex networks.

We now use merge sets, were one task is registered to be merge-able
with multiple tasks. To simplify the code, resolve_input_matching
is now made unaware of the cycle merging, probably at a slight
performance cost.