app: fix delayed startup of the process server
To avoid stalling while the process server starts, syskit is first
starting it in setup and then connecting to it in prepare. However,
in the meantime, we cannot load models from the local process server
(obviously).

Add a fake server (with default_loader as loader) so that we can do
the loading. It gets replaced in connect_to_local_process_server