app: HACK - make roby test --coverage load all task context models
This is a hack: when one runs roby test --coverage, one assumes that
all models should be loaded (so that we get the coverage for all of
them). However, it is harmful in all other app modes (long loading times).

This commit enables load-all if BOTH testing? and auto_load_models? are
set. In the long run, we want to make auto_load_models false by default,
which would allow to handle it gracefully (syskit_load_all? would simply
be replaced by auto_load_models?)