fix callbacks being registered multiple times on Orocos.loader
The cleanup routine was disconnecting from the process servers, but
was actually not really removing them. Moreover, we were not properly
cleaning up after us, in the sense that the on_project_load callback
was not removed from the process servers' loaders.

In load_only_models? mode, this was leading to the callback being
registered as many times as cleanup was called on Orocos.loader, and
therefore to horrible load times.