normalize the selected default objects in DI#add
The normalization (and associated typecheck) was done very late in
resolve_default_selections, with no good reason. Do it early in #add

The check is not done in #add_defaults so that we can bypass
the check when it does not make sense (e.g. when we merge DI objects).
DependencyInjection#add_explicit already behaves this way (associated
check is in #add)