net_gen: remove the removal of cycles in the merge graph
This is code that was there before the cycle merge algorithm got
integrated. It was needed to try very very hard to avoid creating
cycles while merging. Since we do have the cycle merge now, it is
in principle not needed anymore.

It brings the merge algorithm complexity in the worst case of
having to merge the same network multiple times from more than
O(N^2) to something much closer to O(N). However, it does not
really provide a huge performance boost in other cases.