improve performance of resolve_selection_recursively
The method was trying very hard to stick to duck-typing principles,
but was unfortunately doing a lot of work for naught because of this.
Namely, the recursive resolution can only apply to plain models
or bound data services that apply on plain models, but was calling
#to_component_model on e.g. InstanceRequirements or BoundDataService
objects that were pointing to InstanceRequirements.

Restrict the resolution to the object types we actually can do something
about.