runtime: fix critical failure in connection management update in some bad interaction with stopped deployments (closes #2)
If (1) the connection management queues a certain connection for
removal and (2) tasks involving this connection gets killed
(e.g. because the deployment dies) and finally (3) we are unlucky
enough that no other tasks get modified, then there is a conflict
between the cleanup done by Deployment#cleanup_dead_connections and
ConnectionManager#update. This led to having a critical failure
in the disconnection code that led to Roby quitting.

This commit fixes the issue by making sure that #cleanup_dead_connection
removes the relevant entries from DataFlow.pending_changes