net_gen: fix network deployment after 89c4198f9eec
The last step in network deployment is to run a merge pass to merge
the compositions that can be merged together. However,
#merge_identical_tasks now seeds only using task contexts, which
meant that it would not consider the compositions as potential
seeds.

The proper fix is to make #finalize_deployed_tasks seed the merge
solver with the set of merges it did.